services/backend/requirements.txt
django==5.0
gunicorn==21.2.0
psycopg2-binary==2.9.6
djangorestframework==3.14.0
Markdown==3.5.1
django-filter==23.5
drf-spectacular==0.26.5
djoser==2.2.2